IDT fixed BUG, handling of NMI IRQ when FLASH_FLAG_ECCD is SET, now we just clear ECCD flag, we will not use this functionality, AlxMemSafe will handle this ---> also added handling of NOT handled IRQ (default_handler) and also minor visual changes

+++ alxParamStore.c
@@ -47,7 +47,7 @@ void AlxParamStore_Ctor
 //******************************************************************************
 // Functions
 //******************************************************************************
-Alx_Status AlxParamStore_Init(AlxParamStore* me, Alx_Status* statusParamGroup, uint8_t numOfParamGroups)
+Alx_Status AlxParamStore_Init(AlxParamStore* me, Alx_Status* statusParamGroup, uint32_t numOfParamGroups)
 {
 	// Assert
 	ALX_PARAM_STORE_ASSERT(me->wasCtorCalled == true);
